% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Produc.R
\docType{data}
\name{Produc}
\alias{Produc}
\title{Produc: Panel data frame across 48 states. Data are from the Ecdat package. Variables include:}
\format{
An object of class \code{data.frame} with 816 rows and 10 columns.
}
\usage{
data(Produc)
}
\description{
\itemize{
\item \code{state}: state
\item \code{year}: year
\item \code{pcap}: private capital stock
\item \code{hwy}: highway and street
\item \code{water}: water and sewer facilities
\item \code{util}: other public buildings and structures
\item \code{pc}: public capital
\item \code{gsp}: gross state products
\item \code{emp}: labor input measured by the employment in non-agricultural payrolls
\item \code{unemp}: state unemployment rate
}
}
\examples{
data(Produc)
}
\references{
Yves Croissant and Spencer Graves (2020). Ecdat: Data Sets for Econometrics. R package version 0.3-9. https://CRAN.R-project.org/package=Ecdat
}
\keyword{datasets}
