% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaww.R
\docType{data}
\name{usaww}
\alias{usaww}
\title{usaww Spatial weights matrix of the 48 continental united states based on the Queen's neighborhood structure.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 48 rows and 48 columns.
}
\usage{
data(usaww)
}
\description{
usaww Spatial weights matrix of the 48 continental united states based on the Queen's neighborhood structure.
}
\examples{
data(usaww)
}
\references{
Giovanni Millo, Gianfranco Piras (2012). splm: Spatial Panel Data Models in R. Journal of Statistical Software, 47(1), 1-38. URL: http://www.jstatsoft.org/v47/i01/.
}
\keyword{datasets}
