% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlagvars.R
\name{createlagvars}
\alias{createlagvars}
\title{createlagvars}
\usage{
createlagvars(data, vars, id, time = time, wide, ...)
}
\arguments{
\item{data}{Name of dataframe.}

\item{vars}{Response and explanatory variables to be lagged.}

\item{id}{Group identifier (example: state).}

\item{time}{Number of time periods in the dataset. Lags will be taken for each time period. Default is 2 time periods. For a spatial-only regression model, set \code{time=1}.}

\item{wide}{Boolean indicator. Takes \code{TRUE} if data is in wide format and \code{FALSE} if data is in long format. Default is \code{FALSE}.}

\item{...}{Arguments to be passed to \code{dplyr::filter()}}
}
\description{
Creates lagged explanatory and response variables for data in long format.
}
\details{
Create lagged and transformed variables
}
\examples{
\donttest{
 outdf<- createlagvars(data = Produc, vars=c(y,xs), id="state", time=2,wide = FALSE,year==1970 | year==1971)}
}
